move builder class to its own file to improve error reporting
